refactor: move command palette to features
