+++ apps/core/src/features/command-palette/state.ts
@@ -1,0 +1,3 @@
+import { createDialogState } from "@/state/dialogState"
+
+export const useCommandPaletteState = createDialogState()
